refactor: move internal DiscreteInputActionTrigger outside of the public API (yesman.epicfight.api.client.input -> yesman.epicfight.client.input)
